+++ internal/workflow.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/docker/docker/api/types/container"
 	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 	"github.com/tinkerbell/tink/protos/workflow"
@@ -13,6 +14,40 @@ import (
 
 type workflowsFilterByFunc = func(workflow.WorkflowServiceClient, []*workflow.Workflow) []*workflow.Workflow
 type actionsFilterByFunc = func([]*workflow.WorkflowAction) []*workflow.WorkflowAction
+
+// containerConfigOption allows modifying the container config defaults
+type containerConfigOption func(*container.Config)
+
+// containerHostOption allows modifying the container host config defaults
+type containerHostOption func(*container.HostConfig)
+
+// actionToDockerContainerConfig takes a workflowAction and translates it to a docker container config
+func actionToDockerContainerConfig(ctx context.Context, workflowAction workflow.WorkflowAction, opts ...containerConfigOption) *container.Config { // nolint
+	defaultConfig := &container.Config{
+		AttachStdout: true,
+		AttachStderr: true,
+		Tty:          true,
+		Env:          workflowAction.Environment,
+		Cmd:          workflowAction.Command,
+		Image:        workflowAction.Image,
+	}
+	for _, opt := range opts {
+		opt(defaultConfig)
+	}
+	return defaultConfig
+}
+
+func actionToDockerHostConfig(ctx context.Context, workflowAction workflow.WorkflowAction, opts ...containerHostOption) *container.HostConfig { // nolint
+	defaultConfig := &container.HostConfig{
+		Binds:      workflowAction.Volumes,
+		PidMode:    container.PidMode(workflowAction.Pid),
+		Privileged: true,
+	}
+	for _, opt := range opts {
+		opt(defaultConfig)
+	}
+	return defaultConfig
+}
 
 // getAllWorkflows job is to talk with tink server and retrieve all workflows.
 // Optionally, if one or more filterByFunc is passed in,
